Fix crash when a configured monitor is not connected (#51)
If some monitor exists in `config.xml`, but is not connected to the computer,
the control flow will reach MainViewModel.cs:232, where the program tries to
set `newMonitorData.HdrLuts = new ObservableCollection<string>(hdrLutPaths)`.

However, since `<hdr_luts />` may not be present in `config.xml`, `hdrLutPaths`
may be `null`. This will result in a `System.ArgumentNullException` being
thrown during construction of the `ObservableCollection<string>`.

This patch fixes the crash by only setting `newMonitorData.HdrLuts` when the
corresponding `hdrLutPaths` is not `null`.